reworked system how the API works and setup a scheduler
