+++ server/plugins/cronjobs/index.ts
@@ -1,0 +1,16 @@
+import { useScheduler } from "#scheduler";
+import refreshFruchtLaborCache from "./refreshFruchtLaborCache";
+
+export default defineNitroPlugin(() => {
+  startScheduler();
+});
+
+function startScheduler() {
+  const scheduler = useScheduler();
+
+  scheduler
+    .run(() => {
+      refreshFruchtLaborCache();
+    })
+    .dailyAt(4, 0);
+}
